+++ MangaMark_Vue/src/js/website/mangatxWebsite.ts
@@ -1,5 +1,7 @@
+import fetchManga from "../fetchManga";
 import { Manga, initManga } from "../types/manga";
 import getChapterNumber from "../utils/getChapterNumber";
+import handleChapterJump from "../utils/handleChapterJump";
 import Website from "./website";
 
 export default class MangatxWebsite implements Website {
@@ -12,10 +14,19 @@ export default class MangatxWebsite implements Website {
     result.title = fTitleChapter.split("-")[0].trim();
     return result;
   }
-  getMangaOnList(): Manga {
+  async getMangaOnList() {
     const result = { ...initManga };
-    // let fTileChapter = document.querySelectorAll("span[itemprop='name']")
-    // result.title = fTileChapter[2].innerHTML.trim()
+    result.title = document.querySelector<HTMLElement>('h1').innerHTML.trim();
+    const mangaApi = await fetchManga(result, true)
+    if (mangaApi) {
+      const list = document.querySelector('.listing-chapters_wrap > ul')
+      const listItems = list.querySelectorAll('li');
+      for (let i = 0; i < listItems.length; i++) {
+        const li = listItems[i];
+        const a = li.querySelector<HTMLElement>('a');
+        handleChapterJump(a, mangaApi)
+      }
+    }
     return result
   }